+++ client/src/App.js
@@ -14,6 +14,7 @@ import EditProfile from './components/profile-forms/EditProfile';
 import AddRestaurant from './components/profile-forms/AddRestaurant';
 import AddMedia from './components/profile-forms/AddMedia';
 import Lists from './components/lists/Lists';
+import List from './components/list/List';
 
 import PrivateRoute from './components/routing/PrivateRoute';
 
@@ -58,6 +59,9 @@ const App = () => {
                         </Route>
                         <Route exact path='/lists'>
                             <Lists />
+                        </Route>
+                        <Route exact path='/lists/:id'>
+                            <List />
                         </Route>
                         {/* Private Routes */}
                         <PrivateRoute exact path='/dashboard'>